+++ hub_module/modules/video/classification/videotag_tsn_lstm/resource/configs/tsn.yaml
