+++ hub_module/modules/video/classification/videotag_tsn_lstm/resource/models/tsn/tsn.py
